pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "team"
version = "0.0.1"
description = "A package to control an ABB robot running RobotWare >= 7"
authors = [
  { name = "Lorenzo Panchetti", email = "lorenzo.panchetti@schindler.com" },
]
readme = "README.md"
dependencies = [
  "keyboard>=0.13.5",
  "pandas>=1.3.5",
  "dtaidistance>=2.3.6",
  "scipy>=1.7.3",
  "matplotlib>=3.3.4",
  "scikit-learn>=0.24.2",
  "gmr>=1.6.2",
  "numpy>=1.22.1",
  "scikit-optimize@git+https://github.com/MalcolmMielle/scikit-optimize.git",
  "tikzplotlib>=0.10.1",
  "rws2>=0.0.1",
  "spatialmath-python >= 1.1.8"
]

[project.urls]
Home = "https://github.com/SchindlerReGIS/team"
